chore: fix planning page
